// server/src/schemas/auth.schema.ts
import { z } from 'zod'

export const registerSchema = z.object({
	email: z.email(),
	password: z.string().min(8, 'Password must be at least 8 characters'),
	name: z.string().min(1),
})

export const loginSchema = z.object({
	email: z.email(),
	password: z.string().min(1),
})

export type RegisterInput = z.infer<typeof registerSchema>
export type LoginInput = z.infer<typeof loginSchema>